Refactored Entities, added foreign keys. Added redis connection.

+++ src/main/java/com/Cristian/GraduationProject/repository/CostumersRepository.java
@@ -3,5 +3,5 @@ package com.Cristian.GraduationProject.repository;
 import com.Cristian.GraduationProject.entity.Costumers;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CostumersRepository extends JpaRepository<Costumers,Integer> {
+public interface CostumersRepository extends JpaRepository<Costumers,Long> {
 }
